add category to product & edit

+++ database/migrations/2020_09_18_144918_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('price');
+            $table->decimal('price', 10, 1);
             $table->integer('stock');
             $table->text('description');
             $table->integer('is_active')->default(0);
